- addressed code review comments, especially change to partition metadata primary key which required an almost complete change of `LucenePartitioner` and removal of previously added code to `LuceneIndexMaintainer`

+++ fdb-record-layer-lucene/src/test/java/com/apple/foundationdb/record/lucene/LuceneIndexTestUtils.java
@@ -234,7 +234,7 @@ public class LuceneIndexTestUtils {
                 .build();
     }
 
-    public static TestRecordsTextProto.ComplexDocument createComplexDocument(long docId, String text, String text2, long group, int score, boolean isSeen, double time) {
+    public static TestRecordsTextProto.ComplexDocument createComplexDocument(long docId, String text, String text2, long group, int score, boolean isSeen, long time) {
         return TestRecordsTextProto.ComplexDocument.newBuilder()
                 .setDocId(docId)
                 .setText(text)
